fix: espacamento entre cilindros deixa de ser truncado em motores pequenos
grau_entre_cada_cilindro era declarado como byte, e calcular_grau_entre_cada_cilindro
retornava byte, mas o valor passa de 255 em motores de poucos cilindros. Num 4
tempos com o sensor no virabrequim o ciclo tem 720 graus:

  cilindros   correto   armazenado
      1         720        208   <- errado
      2         360        104   <- errado
      3         240        240
      4         180        180
      6         120        120

Com o sensor no comando o ciclo tem 360 graus e so o motor de 1 cilindro
quebra (360 -> 104). O mesmo vale para o fallback de qtd_cilindro < 1.

Esse valor define o espacamento angular entre eventos de cilindros
consecutivos, usado tanto em calcular_angulo_ignicao_indice quanto em
calcular_angulo_injecao_indice - truncado, joga ignicao e injecao em angulos
sem relacao com o motor.

O compilador ja avisava ("large integer implicitly truncated to unsigned
type") no return 360, em toda compilacao.

Nao afeta motores de 3 cilindros ou mais, onde o valor cabe em um byte.
Alargar para int e seguro: todos os usos ja atribuem para int ou multiplicam
em contexto int, e a variavel nao vai para a EEPROM - e sempre recalculada por
calcular_grau_entre_cada_cilindro na carga da configuracao.

Custo no ATmega328P: 28 bytes de flash, 1 de RAM.

+++ include/util.h
@@ -58,15 +58,21 @@ static inline int offset_referencia_roda_fonica_graus() {
 // ciclo completo de combustao: 720 graus (2 voltas) num motor 4 tempos,
 // 360 graus (1 volta) num motor 2 tempos. Com o sensor no comando, o proprio
 // sensor ja gira 1 volta por ciclo, entao sempre 360 graus, nos dois casos.
-static inline byte calcular_grau_entre_cada_cilindro() {
+// Retorna int, nao byte: o resultado passa de 255 em motores de poucos
+// cilindros e era truncado silenciosamente. Num 4 tempos com sensor no
+// virabrequim o ciclo tem 720 graus, entao 2 cilindros dao 360 (virava 104) e
+// 1 cilindro da 720 (virava 208), jogando ignicao e injecao em angulos sem
+// relacao com o motor. O compilador ja avisava disso no "return 360" abaixo.
+// Nao afeta 3 cilindros ou mais, onde o valor cabe em um byte.
+static inline int calcular_grau_entre_cada_cilindro() {
   if (qtd_cilindro < 1) {
     return 360;
   }
   if (local_rodafonica == 2) {
     int graus_ciclo_completo = (tipo_motor == 2) ? 360 : 720;
-    return (byte)(graus_ciclo_completo / qtd_cilindro);
+    return graus_ciclo_completo / qtd_cilindro;
   }
-  return (byte)(360 / qtd_cilindro);
+  return 360 / qtd_cilindro;
 }
 
 float calculateBeta(float ntcResistance1, float ntcTemperature1, float ntcResistance2, float ntcTemperature2) {
